pyproject.toml
[project]
name = "stock-sentiment-analysis"
dynamic = ["version"]
description = "NLP sentiment analysis for securities markets with RoBERTa transformers"
authors = [
    { name = "Nicholas Dykema", email = "nicowork13@gmail.com" },
    { name = "Elliott Steer", email = "essteer@pm.me" },
]
dependencies = [
    "masquer==1.1.0",
    "pandas>=2.2.2",
    "plotly>=5.19.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "spacy>=3.7.4",
    "spacy-transformers>=1.3.4",
    "streamlit>=1.31.1",
    "yfinance>=0.2.37",
]
requires-python = ">=3.9"
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["NLP", "RoBERTa", "sentiment analysis"]
classifiers = [
    "Operating System :: OS Independent",
    "Natural Language :: English",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
documentation = "https://github.com/essteer/stock-sentiment-analysis/blob/main/README.md"
repository = "https://github.com/essteer/stock-sentiment-analysis"
issues = "https://github.com/essteer/stock-sentiment-analysis/issues"
